refactor: rename methods and files
