+++ src/app/common/helpers/date.ts
@@ -1,4 +1,4 @@
-function resetDate(date: Date): Date {
+function startOfDay(date: Date): Date {
   date.setHours(0);
   date.setMinutes(0);
   date.setSeconds(0);
@@ -8,19 +8,19 @@ function resetDate(date: Date): Date {
 
 function getEmptyDate(milliseconds: number): Date {
   const date: Date = new Date(milliseconds);
-  return resetDate(date);
+  return startOfDay(date);
 }
 
 function getNextDay(milliseconds: number): Date {
   const dayMilliseconds: number = 86400000;
   const nextDay: Date = new Date(milliseconds + dayMilliseconds);
-  return resetDate(nextDay);
+  return startOfDay(nextDay);
 }
 
 function getPrevDay(milliseconds: number): Date {
   const dayMilliseconds: number = 86400000;
   const prevDay: Date = new Date(milliseconds - dayMilliseconds);
-  return resetDate(prevDay);
+  return startOfDay(prevDay);
 }
 
 function getClosestEmptyDate<T extends { value: string }>(date: Date, dates: Array<T>): Date {
@@ -42,4 +42,4 @@ function getClosestEmptyDate<T extends { value: string }>(date: Date, dates: Arr
   return emptyDate;
 }
 
-export { resetDate, getNextDay, getPrevDay, getClosestEmptyDate, getEmptyDate };
+export { startOfDay, getNextDay, getPrevDay, getClosestEmptyDate, getEmptyDate };
