+++ src/app/common/models/Table/Table-config.ts
@@ -1,10 +1,10 @@
 import { TableHeaderConfig, Paginator } from './index';
-import { EditMark } from '../mark/edit-mark';
-import { TNullable } from '../useful/tnullable';
+import { EditMark } from '../mark';
+import { TNullable } from '../utils/tnullable';
 
 export interface ITableConfig<T> {
   headers: Array<TableHeaderConfig>;
   body: Array<T>;
   paginator?: TNullable<Paginator>;
-  editCell?: TNullable<EditMark>;
+  editCell?: TNullable<EditMark | undefined>;
 }
