+++ src/app/common/models/Table/Table-header-config.ts
@@ -1,6 +1,6 @@
 import { FormControl } from '@angular/forms';
 
-import { TNullable } from '../useful/tnullable';
+import { TNullable } from '../utils/tnullable';
 
 interface ITableHeaderConfig {
   value: TNullable<string>;
