+++ src/app/common/models/utils/key-check.ts
