+++ src/app/common/services/confirm-save.service.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core';
 import { Observable, of } from 'rxjs';
 import { map } from 'rxjs/operators';
 
-import { IConfirmSave } from '../models/useful/confirm-save';
+import { IConfirmSave } from '../models/utils/confirm-save';
 
 @Injectable()
 export class ConfirmSaveService<T> {
