+++ src/app/students/student-form/student-form.component.ts
@@ -9,7 +9,7 @@ import { Student } from '../../common/models/student';
 import { StudentService } from '../services/student.service';
 import { FormComponent } from '../../shared/components';
 import { ConfirmSaveService } from '../../common/services';
-import { IConfirmSave } from '../../common/models/useful/confirm-save';
+import { IConfirmSave } from '../../common/models/utils/confirm-save';
 
 @Component({
   selector: 'app-student-form',
