+++ src/app/subjects/services/table-config-history.service.ts
@@ -1,12 +1,12 @@
 import { Injectable } from '@angular/core';
 
 import { convert } from '../helpers/convert-history-change';
-import { DateChanges } from '../../common/models/useful/date-changes';
+import { DateChanges } from '../../common/models/utils/date-changes';
 import { TableConfigHistory } from '../../common/models/table/table-config-history';
 import { Mark, MarkHistory, IMarksByDate } from '../../common/models/mark';
-import { IChangeField } from '../../common/models/table/table-change-field';
-import { IDataChanges } from '../../common/models/useful/data-changes';
-import { TNullable } from 'src/app/common/models/useful/tnullable';
+import { IChangeField } from '../../common/models/table';
+import { IDataChanges } from '../../common/models/utils/data-changes';
+import { TNullable } from 'src/app/common/models/utils/tnullable';
 
 @Injectable()
 export class TableConfigHistoryService {
