+++ src/app/url.ts
@@ -1,4 +1,4 @@
-import { IUrl } from './common/models/useful/url';
+import { IUrl } from './common/models/utils/url';
 
 export const urlProvider: IUrl = {
   students: 'http://localhost:3001/students',
